下车，完成订单

+++ QiPinCheClient/EGORefresh/RefreshTableView.m
@@ -111,7 +111,7 @@
 //	else {
 //		[self delConfirmView];
 //	}
-	int count = [self.tableArray count];
+	NSUInteger count = [self.tableArray count];
 	return count;
 }
